replace pcre with pcre2
Since pcre is unmaintained, the usage of ocaml-pcre will be replaced
with ocaml-pcre2. While other regular expession libraries for OCaml
exists, they are not as feature complete as pcre2 is.

This is the minimal change to use the system ocaml-pcre2. To reduce the
size of this patch, a separate commit might be required to import
pcre2-ocaml.git into bundles and use it in the build system.

Signed-off-by: Olaf Hering <olaf@aepfle.de>